requirements.txt
fastapi
uvicorn
mysql-connector-python
pdfminer.six
pytesseract
requests
watchdog
python-multipart
jinja2
llama-cpp-python
